Se añaden endpoints para administradores

+++ src/client/application/delete/index.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
+import { RentRepository } from '../../../rent/infrastructure/persistence/repository/rent.repository';
 import { KeycloakRepository } from '../../infrastructure/idp/keycloak/repositories/keycloak.repository';
 import { ClientRepository } from '../../infrastructure/persistence/repository/client.repository';
 
@@ -7,6 +8,7 @@ export class DeleteClientService {
   constructor(
     private readonly clientRepository: ClientRepository,
     private readonly keycloakRepository: KeycloakRepository,
+    private readonly rentRepository: RentRepository,
   ) {}
 
   async delete(id: string, email: string) {
@@ -15,7 +17,22 @@ export class DeleteClientService {
         email,
       },
     });
-    await this.keycloakRepository.deleteUser(id);
+    try {
+      await this.keycloakRepository.deleteUser(id);
+    } catch (e) {
+      console.error(e);
+    }
+    const rent = await this.rentRepository.find({
+      status: 'checkedin',
+      renterUser: {
+        id,
+      },
+    });
+
+    if (rent.length > 0) {
+      throw new BadRequestException('User has a checked rent in');
+    }
+
     return this.clientRepository.delete(user.id);
   }
 }
